2dcontext/fill-and-stroke-styles/2d.pattern.image.incomplete.removedsrc.html

<!DOCTYPE html>
<title>Canvas test: 2d.pattern.image.incomplete.removedsrc</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../../../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.pattern.html">pattern</a>.<a href="index.2d.pattern.image.html">image</a>.<a href="index.2d.pattern.image.incomplete.html">incomplete</a>.removedsrc</h1>
<p class="desc"></p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.pattern.incomplete.image">2d.pattern.incomplete.image</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

var img = document.getElementById('red.png');
img.removeAttribute('src');
_assertSame(ctx.createPattern(img, 'repeat'), null, "ctx.createPattern(img, 'repeat')", "null");


});
</script>
<img src="../../../images/red.png" id="red.png" class="resource">
